feat: fetchPlaceholders project version (#92)
* feat: fetchPlaceholders project version

* chore: align aem.js with latest version
