+++ blocks/search/search.js
@@ -1,8 +1,8 @@
 import {
   createOptimizedPicture,
   decorateIcons,
-  fetchPlaceholders,
 } from '../../scripts/aem.js';
+import { fetchPlaceholders } from '../../scripts/placeholders.js';
 
 const searchParams = new URLSearchParams(window.location.search);
